+++ src/pages/DocumentDisplay.vue
@@ -834,11 +834,12 @@ export default class PageDocumentDisplay extends BaseClass {
     if (!value ||
     (Array.isArray(value) && value.length === 0) ||
     // @ts-ignore
-     (value?.value && value.value.length === 0) ||
+     (value?.value?.length === 0) ||
     // @ts-ignore
      (value.value === null)) {
       return false
     }
+
     return true
   }